toolkit/components/passwordmgr/test/mochitest/test_autocomplete_password_open.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test password field autocomplete footer with and without logins</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="../../../satchel/test/satchel_common.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>

<div id="content"></div>

<pre id="test">
Test for Login Manager: Test password field autocomplete footer with and without logins

<template id="form1-template">
  <form id="form1" action="https://autocomplete:8888/formtest.js">
    <input type="text" name="uname">
    <input type="password" name="pword">
  </form>
</template>

<script class="testbody" type="text/javascript">
  const formTemplate = document.getElementById("form1-template");

  add_named_task("no autofill", async () => {
    await setStoredLoginsDuringTask(
      [location.origin, "", null, "user-1", "pass-1", "uname", "pword"],
      [location.origin, "", null, "user-2", "pass-2", "uname", "pword"]
    );
    const form = setContentForTask(formTemplate);
    const autofillResult = await formAutofillResult(form.id);
    is(autofillResult, "multiple_logins", "form has not been filled due to multiple logins");

    // Make sure initial form is empty as autofill shouldn't happen in the sandboxed frame.
    is(form.uname.value, "", "username is empty");
    is(form.pword.value, "", "password is empty");

    const popupState = await getPopupState();
    is(popupState.open, false, "Check popup is initially closed");
  });

  add_named_task("two logins", async () => {
    await setStoredLoginsDuringTask(
      [location.origin, "", null, "user-1", "pass-1", "uname", "pword"],
      [location.origin, "", null, "user-2", "pass-2", "uname", "pword"]
    );
    const form = setContentForTask(formTemplate);
    const autofillResult = await formAutofillResult(form.id);
    is(autofillResult, "multiple_logins", "form has not been filled due to multiple logins");

    await popupBy(() => form.uname.focus());

    // popup on the password field should open upon focus
    const results = await popupBy(() => synthesizeKey("KEY_Tab"));

    const popupState = await getPopupState();
    is(popupState.selectedIndex, -1, "Check no entries are selected upon opening");

    const expectedMenuItems = [
      "user-1",
      "user-2",
    ];
    checkAutoCompleteResults(results, expectedMenuItems, window.location.host, "Check all menuitems are displayed correctly.");

    is(form.uname.value, "", "username is empty");
    is(form.pword.value, "", "password is empty");
  });

  add_named_task("zero logins", async () => {
    // no logins stored
    await setStoredLoginsDuringTask();

    const form = setContentForTask(formTemplate);
    const autofillResult = await formAutofillResult(form.id);
    is(autofillResult, "no_saved_logins", "form has not been filled due to no saved logins");

    form.uname.focus();

    const shownPromise = popupBy().then(() => ok(false, "Should not have shown"));
    // Popup on the password field should NOT automatically open upon focus when there are no saved logins.
    synthesizeKey("KEY_Tab"); // focus the password field
    SimpleTest.requestFlakyTimeout("Giving a chance for the unexpected popup to show");
    await Promise.race([
      shownPromise,
      new Promise(resolve => setTimeout(resolve, 2000)), // Wait 2s for the popup to appear
    ]);

    const popupState = await getPopupState();
    is(popupState.open, false, "Check popup is still closed");

    is(form.uname.value, "", "username is empty");
    is(form.pword.value, "", "password is empty");

    info("arrow down should still open the popup");
    const autocompleteItems1 = await popupByArrowDown();
    checkAutoCompleteResults(autocompleteItems1, [], window.location.host, "Check only footer is displayed.");
    is(form.uname.value, "", "username is empty");
    is(form.pword.value, "", "password is empty");
  });
</script>
</pre>
</body>
</html>
